==> data_analysis/Analysis_3_post_analysis.R <==
library(data.table); library(tidyverse); library(ggplot2)
df_plt1 = fread("~/Dropbox/Lin/FusioMR/af_predict/r1_analysis1/dfclump_ldl_ais_eur_sas.csv")
df_plt1 = df_plt1 %>% mutate(min_exp = exp2.beta - 1.96*exp2.se, max_exp = exp2.beta + 1.96*exp2.se, min_out = out2.beta - 1.96*out2.se, max_out = out2.beta + 1.96*out2.se) %>% mutate(min_out = ifelse(min_out < -0.3, -0.3, min_out), max_out = ifelse(max_out > 0.4, 0.4, max_out))
pp1 = ggplot(df_plt1, aes(exp2.beta, out2.beta)) + geom_errorbar(aes(ymin = min_out, ymax = max_out), width = 0.001, col = 'gray') + geom_errorbarh(aes(xmin = min_exp, xmax = max_exp), height = 0.001, col = 'gray') + geom_point(col = 'blue', alpha = 0.5, size = 1) + theme_bw() + labs(title = expression("Estimated causal effect of LDL on IS in SAS by FusioMR"["m"]), y = "GWAS effect size for IS", x = "GWAS effect size for LDL") + geom_abline(slope = 0.30, intercept = 0) + theme(axis.title = element_text(size = 14), axis.title.x = element_text(size = 14), axis.title.y = element_text(size = 14), panel.grid.major = element_blank(), panel.grid.minor = element_blank()) + ylim(-0.3,0.4)
df_plt2 = fread("~/Dropbox/Lin/FusioMR/af_predict/r1_analysis1/dfclump_ldl_ais_sas_1e-04.csv")
pp2 = ggplot(df_plt2, aes(exp2.beta, out2.beta, col = is_shared)) + geom_errorbar(aes(ymin = out2.beta - 1.96*out2.se, ymax = out2.beta + 1.96*out2.se), width = 0.001, col = 'gray') + geom_errorbarh(aes(xmin = exp2.beta - 1.96*exp2.se, xmax = exp2.beta + 1.96*exp2.se), height = 0.001, col = 'gray') + geom_point(col = 'blue', alpha = 0.5, size = 1) + labs(title = expression("Estimated causal effect of LDL on IS in SAS by FusioMR"["s"]), y = "GWAS effect size for IS", x = "GWAS effect size for LDL") + theme_bw() + geom_abline(slope = 0.19, intercept = 0) + theme(axis.title = element_text(size = 14), axis.title.x = element_text(size = 14), axis.title.y = element_text(size = 14), panel.grid.major = element_blank(), panel.grid.minor = element_blank())

plot(pp1)
plot(pp2)
